+++ addon/-private/auth/methods/registry.js
